Add header to handle exported objects visibility
There are differences on how different compilers handle the
visibility of exported objects when building a library, so since
we are targeting multi-compiler/platform support, we need to be
able to handle these differences.

The main culprit of this is MSVC, since Clang and gcc export all
objects as a public API meanwhile you have to explicitly export
each object with a macro that's Windows/MSVC only.